move useRowComponent hook to seperate file
